Halfway done with new html renderer.
