+++ dev/Core/UI/HTML/Elements/AElement.cs
@@ -20,6 +20,9 @@ namespace UltimaXNA.Core.UI.HTML.Elements
         public abstract int Width { get; set; }
         public abstract int Height { get; set; }
 
+        public int Layout_X = 0;
+        public int Layout_Y = 0;
+
         public bool CanBreakAtThisAtom
         {
             get
